Changed: unitId Attribute from models
Made unitId from the model classes a long instead of an int.

+++ DataFlowWebservice/DataFlowWebservice/Models/Event.cs
@@ -10,7 +10,7 @@ namespace DataFlowWebservice.Models
     {
         public ObjectId _id { get; set; }
         public DateTime dateTime { get; set; }
-        public int unitId { get; set; }
+        public long unitId { get; set; }
         public string port { get; set; }
         public int portValue { get; set; }
     }
