+++ DataFlowWebservice/DataFlowWebservice/Models/Monitoring.cs
@@ -9,7 +9,7 @@ namespace DataFlowWebservice.Models
     public class Monitoring : IResponseModel
     {
         public ObjectId _id { get; set; }
-        public int unitId { get; set; }
+        public long unitId { get; set; }
         public DateTime beginTime { get; set; }
         public DateTime endTime { get; set; }
         public string sensorType { get; set; }
